Makefile is now working after calling make clean!
The problem was I was using immediate expansion for the ldobjects variable at the top of the file. This caused the first run of make to expand immediately a wildcard that hit no objects because none had been generated yet. The folder names could be better but we will work on more organization later. The nice thing is we now have an expandable framework.